fix :: 이미지 파일 확장자 vadlidation

+++ src/main/kotlin/com/dotori/v2/domain/member/service/impl/UpdateProfileImageServiceImpl.kt
@@ -1,5 +1,6 @@
 package com.dotori.v2.domain.member.service.impl
 
+import com.dotori.v2.domain.member.exception.NotAcceptImgExtensionException
 import com.dotori.v2.domain.member.service.UpdateProfileImageService
 import com.dotori.v2.global.config.redis.service.RedisCacheService
 import com.dotori.v2.global.thirdparty.aws.s3.S3Service
@@ -16,6 +17,14 @@ class UpdateProfileImageServiceImpl(
     private val redisCacheService: RedisCacheService
 ): UpdateProfileImageService {
     override fun execute(multipartFiles: MultipartFile?) {
+
+        val acceptList = listOf("jpg", "jpeg", "png")
+        val splitFile = multipartFiles?.originalFilename.toString().split(".")
+        val extension = splitFile.last().lowercase()
+
+        if (acceptList.none { it == extension })
+            throw NotAcceptImgExtensionException()
+
         val member = userUtil.fetchCurrentUser()
         var uploadFile: String? = s3Service.uploadSingleFile(multipartFiles)
         s3Service.deleteFile(member.profileImage!!)
